Kreirana AuthController i rute za autentifikaciju. Izmenjeni ostali kontroleri

+++ prvi_domaci/app/Http/Controllers/CompanyController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Company;
+
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CompanyController extends Controller
 {
@@ -34,8 +36,9 @@ class CompanyController extends Controller
             "description"=>"nullable|string|max:255",
             "address"=>"nullable|string|max:255",
             "phone"=>"nullable|string|max:20",
-            "user_id"=>"required|exists:users,id"
+            
         ]);
+        $fields['user_id']=Auth::id();
         $company = Company::create($fields);
         return response()->json([
             'message' => 'Company created successfully.',
@@ -73,11 +76,13 @@ class CompanyController extends Controller
             "description"=>"nullable|string|max:255",
             "address"=>"nullable|string|max:255",
             "phone"=>"nullable|string|max:20",
-            "user_id"=>"required|exists:users,id"
         ]);
         $company = Company::find($id);
         if(is_null($company)){
             return response()->json("Company not found.",404);
+        }
+        if ($company->user_id !== Auth::id()) {
+            return response()->json(['message' => 'You cannot update this company because you are not the owner..'], 403);
         }
        $company->update($validated);
        return response()->json(['message' => 'Company updated successfully', 'company' => $company], 200);
@@ -88,6 +93,9 @@ class CompanyController extends Controller
      */
     public function destroy(Company $company)
     {
+        if(Auth::id()!==$company->user_id){
+            return response()->json(["message"=>"You cannot delete this company because you are not the owner."],403);
+        }
         $company->delete();
         return response()->json(['message' => 'Company deleted successfully'], 200);
     }
